Split docutils/sphinx fixture testing
